Реалізовано прив'язку покупок до певного користувача

+++ src/main/java/com/purchasehistorysystem/repository/CategoryRepository.java
@@ -8,22 +8,24 @@ import java.util.List;
 import java.sql.*;
 
 public class CategoryRepository {
-    public List<Category> getAllCategories() throws SQLException {
+    public List<Category> getAllCategories(int userId) throws SQLException {
         List<Category> categories = new ArrayList<>();
 
-        categories.add(new Category(0, "Всі", null));
+        categories.add(new Category(0, "Всі", null, 0));
 
-        String sqlQuery = "SELECT * FROM categories";
+        String sqlQuery = "SELECT * FROM categories WHERE user_id = ?";
 
         try (
                 Connection connection = Database.databaseConnection();
-                Statement statement = connection.createStatement();
-                ResultSet resultSet = statement.executeQuery(sqlQuery)
+                PreparedStatement preparedStatement = connection.prepareStatement(sqlQuery);
                 ) {
+            preparedStatement.setInt(1, userId);
+            ResultSet resultSet = preparedStatement.executeQuery();
             while (resultSet.next()) {
                 Category category = new Category(resultSet.getInt("id"),
                         resultSet.getString("name"),
-                        resultSet.getString("icon_path"));
+                        resultSet.getString("icon_path"),
+                        resultSet.getInt("user_id"));
 
                 categories.add(category);
             }
@@ -31,8 +33,8 @@ public class CategoryRepository {
         return categories;
     }
 
-    public void addCategory(Category category) throws  SQLException {
-        String sqlQuery = "INSERT INTO categories (name, icon_path) VALUES (?, ?)";
+    public void addCategory(Category category, int userId) throws  SQLException {
+        String sqlQuery = "INSERT INTO categories (name, icon_path, user_id) VALUES (?, ?, ?)";
 
         try (
                 Connection connection = Database.databaseConnection();
@@ -40,18 +42,20 @@ public class CategoryRepository {
                 ) {
             preparedStatement.setString(1, category.getName());
             preparedStatement.setString(2, category.getIconPath());
+            preparedStatement.setInt(3, userId);
             preparedStatement.executeUpdate();
         }
     }
 
-    public void deleteCategory(int id) throws SQLException {
-        String sqlQuery = "DELETE FROM categories WHERE id = ?";
+    public void deleteCategory(int id, int userId) throws SQLException {
+        String sqlQuery = "DELETE FROM categories WHERE id = ? AND user_id = ?";
 
         try (
                 Connection connection = Database.databaseConnection();
                 PreparedStatement preparedStatement = connection.prepareStatement(sqlQuery);
                 ) {
             preparedStatement.setInt(1, id);
+            preparedStatement.setInt(2, userId);
             preparedStatement.executeUpdate();
         }
     }
